requirements.txt
requests
python-dotenv
jieba
unidecode
mysql-connector-python
openai
flask